requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.54.0
djangorestframework==3.14.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.6
Pillow==9.5.0
pycparser==2.21
PyJWT==2.7.0
pypng==0.20220715.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
qrcode==7.4.2
requests==2.30.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
yarl==1.9.2
